Разделил по классам файлы, добавил NetworkConnectionProvider, изменил имена файлов на dataSource, убрал прокидывание errorCode тк это неинформативно, добавил несколько error enum в разных слоях, добавил несколько констант для того чтобы избавиться от магических цифр
