+++ app/src/main/java/ru/practicum/android/diploma/common/domain/model/vacancy_models/Experience.kt
@@ -1,0 +1,6 @@
+package ru.practicum.android.diploma.common.domain.model.vacancy_models
+
+data class Experience(
+    val id: String, // требуемый опыт работы из справочника experience: "between1And3"
+    val name: String // название опыта работы: "От 1 года до 3 лет"
+)
