+++ app/src/main/java/ru/practicum/android/diploma/common/domain/model/vacancy_models/KeySkill.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.common.domain.model.vacancy_models
+
+data class KeySkill(
+    val name: String // название ключевого навыка: "Первичный документооборот"
+)
