+++ app/src/main/java/ru/practicum/android/diploma/search/data/dataSource/VacancyRemoteDataSource.kt
@@ -1,0 +1,7 @@
+package ru.practicum.android.diploma.search.data.dataSource
+
+import ru.practicum.android.diploma.search.data.network.Response
+
+interface VacancyRemoteDataSource {
+    suspend fun doRequest(dto: Any): Response
+}
