+++ app/src/main/java/ru/practicum/android/diploma/search/data/model/ErrorRemoteDataSource.kt
@@ -1,0 +1,6 @@
+package ru.practicum.android.diploma.search.data.model
+
+enum class ErrorRemoteDataSource {
+    NO_CONNECTION,
+    ERROR_OCCURRED
+}
