+++ app/src/main/java/ru/practicum/android/diploma/search/data/model/SearchRequest.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.search.data.model.dto
+package ru.practicum.android.diploma.search.data.model
 
 // отдельный класс для каждого запроса, поля которого
 // все параметры для соответствующего запроса, даже если такой параметр всего один
