+++ app/src/main/java/ru/practicum/android/diploma/search/data/model/dto/EmploymentDto.kt
@@ -1,0 +1,6 @@
+package ru.practicum.android.diploma.search.data.model.dto
+
+data class EmploymentDto(
+    val id: String, // тип занятости из справочника employment: "full"
+    val name: String // название типа занятости: "Полная занятость"
+)
