+++ app/src/main/java/ru/practicum/android/diploma/search/data/model/dto/ExperienceDto.kt
@@ -1,0 +1,6 @@
+package ru.practicum.android.diploma.search.data.model.dto
+
+data class ExperienceDto(
+    val id: String, // требуемый опыт работы из справочника experience: "between1And3"
+    val name: String // название опыта работы: "От 1 года до 3 лет"
+)
