+++ app/src/main/java/ru/practicum/android/diploma/search/data/model/dto/KeySkillDto.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.search.data.model.dto
+
+data class KeySkillDto(
+    val name: String // название ключевого навыка: "Первичный документооборот"
+)
