+++ app/src/main/java/ru/practicum/android/diploma/search/data/model/dto/SalaryDto.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.search.data.model.dto
+
+data class SalaryDto(
+    val currency: String?, // Код валюты из справочника currency: "RUR",
+    val from: Int?, // Нижняя граница зарплаты: 30000,
+    val gross: Boolean?, // Признак что границы зарплаты указаны до вычета налогов: true,
+    val to: Int? // Верхняя граница зарплаты: null
+)
