+++ app/src/main/java/ru/practicum/android/diploma/search/data/model/dto/ScheduleDto.kt
@@ -1,0 +1,6 @@
+package ru.practicum.android.diploma.search.data.model.dto
+
+data class ScheduleDto(
+    val id: String, // график работы из справочника schedule: "fullDay"
+    val name: String // Название графика работы: "Полный день" | "Удаленная работа"
+)
