+++ app/src/main/java/ru/practicum/android/diploma/search/data/network/HeadHunterApiService.kt
@@ -2,7 +2,7 @@ package ru.practicum.android.diploma.search.data.network
 
 import retrofit2.http.GET
 import retrofit2.http.Query
-import ru.practicum.android.diploma.search.data.model.dto.SearchResponse
+import ru.practicum.android.diploma.search.data.model.SearchResponse
 
 interface HeadHunterApiService {
     @GET("/vacancies")
