--- app/src/main/java/ru/practicum/android/diploma/search/data/network/RetrofitNetworkClient.kt
@@ -1,51 +1,0 @@
-package ru.practicum.android.diploma.search.data.network
-
-import android.content.Context
-import android.net.ConnectivityManager
-import android.net.NetworkCapabilities
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-import ru.practicum.android.diploma.search.data.model.dto.SearchRequest
-
-// Retrofit - реализация интерфейса NetworkClient
-class RetrofitNetworkClient(
-    private val headHunterApiService: HeadHunterApiService,
-    private val context: Context
-) : NetworkClient {
-
-    override suspend fun doRequest(dto: Any): Response {
-
-        if (!isConnected()) {
-            return Response().apply { resultCode = -1 }
-        }
-        if (dto !is SearchRequest) {
-            return Response().apply { resultCode = 400 }
-        }
-
-        return withContext(Dispatchers.IO) {
-            try {
-                val response = headHunterApiService.search(dto.text)
-                response.apply { resultCode = 200 }
-            } catch (e: Throwable) {
-                Response().apply { resultCode = 500 }
-            }
-        }
-
-    }
-
-    private fun isConnected(): Boolean {
-        val connectivityManager = context.getSystemService(
-            Context.CONNECTIVITY_SERVICE
-        ) as ConnectivityManager
-        val capabilities =
-            connectivityManager.getNetworkCapabilities(connectivityManager.activeNetwork)
-        if (capabilities != null) {
-            when {
-                capabilities.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) -> return true
-                capabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) -> return true
-                capabilities.hasTransport(NetworkCapabilities.TRANSPORT_ETHERNET) -> return true
-            }
-        }
-        return false
-    }
-}
