+++ app/src/main/java/ru/practicum/android/diploma/search/domain/model/ErrorStatusDomain.kt
@@ -1,0 +1,6 @@
+package ru.practicum.android.diploma.search.domain.model
+
+enum class ErrorStatusDomain {
+    NO_CONNECTION,
+    ERROR_OCCURRED
+}
