+++ app/src/main/java/ru/practicum/android/diploma/search/domain/repository/NetworkConnectionProvider.kt
@@ -1,0 +1,5 @@
+package ru.practicum.android.diploma.search.domain.repository
+
+interface NetworkConnectionProvider {
+    fun isConnected(): Boolean
+}
