+++ app/src/main/java/ru/practicum/android/diploma/search/domain/useCase/SearchUseCase.kt
@@ -1,8 +1,9 @@
 package ru.practicum.android.diploma.search.domain.useCase
 
 import kotlinx.coroutines.flow.Flow
-import ru.practicum.android.diploma.common.domain.models.Vacancy
+import ru.practicum.android.diploma.common.domain.model.vacancy_models.Vacancy
+import ru.practicum.android.diploma.search.domain.model.ErrorStatusDomain
 
 interface SearchUseCase {
-    fun search(expression: String): Flow<Pair<List<Vacancy>?, Int?>>
+    fun search(expression: String): Flow<Pair<List<Vacancy>?, ErrorStatusDomain?>>
 }
