+++ app/src/main/java/ru/practicum/android/diploma/search/domain/useCase/SearchUseCaseImpl.kt
@@ -1,0 +1,30 @@
+package ru.practicum.android.diploma.search.domain.useCase
+
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.map
+import ru.practicum.android.diploma.common.domain.model.vacancy_models.Vacancy
+import ru.practicum.android.diploma.common.util.Resource
+import ru.practicum.android.diploma.search.data.model.ErrorRemoteDataSource
+import ru.practicum.android.diploma.search.domain.model.ErrorStatusDomain
+import ru.practicum.android.diploma.search.domain.repository.SearchRepository
+
+class SearchUseCaseImpl(private val repository: SearchRepository) : SearchUseCase {
+    override fun search(expression: String): Flow<Pair<List<Vacancy>?, ErrorStatusDomain?>> {
+        return repository.search(expression).map { result ->
+            when (result) {
+                is Resource.Success -> {
+                    Pair(result.data, null)
+                }
+
+                is Resource.Error -> {
+                    val errorStatusDomain = when (result.errorStatus) {
+                        ErrorRemoteDataSource.NO_CONNECTION -> ErrorStatusDomain.NO_CONNECTION
+                        ErrorRemoteDataSource.ERROR_OCCURRED -> ErrorStatusDomain.ERROR_OCCURRED
+                        else -> null
+                    }
+                    Pair(null, errorStatusDomain)
+                }
+            }
+        }
+    }
+}
